This program is used for clustering
-f to select files (npz) for clustering
-out to select where to put the output
-dt sets the distance threshold
-c sets the clustering algorithm; 1 for DBSCAN and 2 for Agglomerative Clustering  By default, DBSCAN is used.

DBSCAN creates clusters based on if each point is within the distance threshold of any point in a group.  This is the one that was initially desired.
Agglomerative Clustering picks a point and adds points to the cluster if it is within the distance threshold of the chosen point.  Not the clustering algorithm initially desired, but it is included just in case it is needed later